put server cmd into main and put redis cmd into persistence/redis.go.
