+++ flow-store/persistence/redis.go
@@ -1,15 +1,14 @@
-package cmd
+package persistence
 
 import (
+	"gopkg.in/redis.v4"
 	"flag"
-	"github.com/op/go-logging"
 )
 
-var log = logging.MustGetLogger("cmd")
-var Port = flag.String("p", "9098", "The port of the server")
+var ReditClient = redis.NewClient(&redis.Options{Addr: *RedisAddr, Password:"", })
+
 var RedisAddr = flag.String("addr", "localhost:6379", "The address of redis server, for example: localhost:6379")
 
 func init() {
-	log.Info("Starting parse cmd line paramters")
 	flag.Parse() // get the arguments from command line
 }
